+++ core/src/com/unciv/logic/civilization/CivInfoStats.kt
@@ -6,6 +6,8 @@ import com.unciv.models.ruleset.tile.ResourceType
 import com.unciv.models.stats.Stat
 import com.unciv.models.stats.StatMap
 import com.unciv.models.stats.Stats
+import com.unciv.models.translations.getPlaceholderParameters
+import com.unciv.models.translations.getPlaceholderText
 import kotlin.math.max
 import kotlin.math.min
 import kotlin.math.pow
@@ -80,16 +82,35 @@ class CivInfoStats(val civInfo: CivilizationInfo) {
                 statMap.add(entry.key, entry.value)
         }
 
-        //City-States culture bonus
+        //City-States bonuses
         for (otherCiv in civInfo.getKnownCivs()) {
-            if (otherCiv.isCityState() && otherCiv.cityStateType == CityStateType.Cultured
-                    && otherCiv.getDiplomacyManager(civInfo.civName).relationshipLevel() >= RelationshipLevel.Friend) {
-                val cultureBonus = Stats()
-                var culture = 3f * (civInfo.getEraNumber() + 1)
+            if (otherCiv.isCityState() && otherCiv.getDiplomacyManager(civInfo.civName).relationshipLevel() >= RelationshipLevel.Friend) {
+                val cityStateBonus = Stats()
+
+                val relevantBonuses =   if (otherCiv.getDiplomacyManager(civInfo.civName).relationshipLevel() == RelationshipLevel.Friend)
+                                            civInfo.getEraObject().friendBonus[otherCiv.cityStateType.name]
+                                        else
+                                            civInfo.getEraObject().allyBonus[otherCiv.cityStateType.name]
+
+                if (relevantBonuses != null) {
+                    for (bonus in relevantBonuses) {
+                        if (bonus.getPlaceholderText() == "Provides [] [] per turn") {
+                            cityStateBonus.add(Stat.valueOf(bonus.getPlaceholderParameters()[1]), bonus.getPlaceholderParameters()[0].toFloat())
+                        }
+                    }
+                } else {
+                    // Deprecated, assume Civ V values for compatibility
+                    if (otherCiv.cityStateType == CityStateType.Cultured) {
+                        cityStateBonus.culture = if(civInfo.getEraNumber() in 0..1) 3f else if (civInfo.getEraNumber() in 2..3) 6f else 13f
+                        if (otherCiv.getDiplomacyManager(civInfo.civName).relationshipLevel() == RelationshipLevel.Ally)
+                            cityStateBonus.culture *= 2f
+                    }
+                }
+
                 if (civInfo.hasUnique("Food and Culture from Friendly City-States are increased by 50%"))
-                    culture *= 1.5f
-                cultureBonus.add(Stat.Culture, culture)
-                statMap.add("City-States", cultureBonus)
+                    cityStateBonus.culture *= 1.5f
+
+                statMap.add("City-States", cityStateBonus)
             }
 
             if (otherCiv.isCityState())
@@ -107,6 +128,18 @@ class CivInfoStats(val civInfo: CivilizationInfo) {
 
         statMap["Transportation upkeep"] = Stats().apply { gold = -getTransportationUpkeep().toFloat() }
         statMap["Unit upkeep"] = Stats().apply { gold = -getUnitMaintenance().toFloat() }
+
+        if (civInfo.religionManager.religion != null) {
+            for (unique in civInfo.religionManager.religion!!.getFounderBeliefs().flatMap { it.uniqueObjects }) {
+                if (unique.placeholderText == "[] for each global city following this religion") {
+                    statMap.add(
+                        "Religion", 
+                        unique.stats.times(civInfo.religionManager.numberOfCitiesFollowingThisReligion().toFloat())
+                    )
+                }
+            }
+        }
+        
 
         if (civInfo.hasUnique("50% of excess happiness added to culture towards policies")) {
             val happiness = civInfo.getHappiness()
@@ -149,13 +182,14 @@ class CivInfoStats(val civInfo: CivilizationInfo) {
         val luxuriesProvidedByCityStates = 
             civInfo.getKnownCivs().asSequence()
                 .filter { it.isCityState() && it.getAllyCiv() == civInfo.civName }
-                .map { it.getCivResources().map { res -> res.resource } }
-                .flatten().distinct().count { it.resourceType === ResourceType.Luxury }
+                .flatMap { it.getCivResources().map { res -> res.resource } }
+                .distinct().count { it.resourceType === ResourceType.Luxury }
         
         statMap["City-State Luxuries"] = happinessBonusForCityStateProvidedLuxuries * luxuriesProvidedByCityStates * happinessPerUniqueLuxury
 
         val luxuriesAllOfWhichAreTradedAway = civInfo.detailedCivResources
-            .filter { it.amount < 0 && it.resource.resourceType == ResourceType.Luxury && (it.origin == "Trade" || it.origin == "Trade request")}
+            .filter { it.amount < 0 && it.resource.resourceType == ResourceType.Luxury
+                    && (it.origin == "Trade" || it.origin == "Trade request")}
             .map { it.resource }
             .filter { !ownedLuxuries.contains(it) }
         
@@ -185,14 +219,46 @@ class CivInfoStats(val civInfo: CivilizationInfo) {
 
         statMap["Natural Wonders"] = happinessPerNaturalWonder * civInfo.naturalWonders.size
 
+        if (civInfo.religionManager.religion != null) {
+            statMap["Religion"] = 0f
+            for (unique in civInfo.religionManager.religion!!.getFounderBeliefs().flatMap { it.uniqueObjects }) {
+                if (unique.placeholderText == "[] for each global city following this religion") {
+                    statMap["Religion"] = 
+                        statMap["Religion"]!! +
+                        unique.stats.happiness * civInfo.religionManager.numberOfCitiesFollowingThisReligion().toFloat()
+                }
+            }
+            if (statMap["Religion"] == 0f) 
+                statMap.remove("Religion")
+        }
+        
         //From city-states
         for (otherCiv in civInfo.getKnownCivs()) {
-            if (otherCiv.isCityState() && otherCiv.cityStateType == CityStateType.Mercantile
-                    && otherCiv.getDiplomacyManager(civInfo).relationshipLevel() >= RelationshipLevel.Friend) {
-                if (statMap.containsKey("City-States"))
-                    statMap["City-States"] = statMap["City-States"]!! + 3f
-                else
-                    statMap["City-States"] = 3f
+            if (otherCiv.isCityState() && otherCiv.getDiplomacyManager(civInfo).relationshipLevel() >= RelationshipLevel.Friend) {
+                val relevantbonuses = if (otherCiv.getDiplomacyManager(civInfo).relationshipLevel() == RelationshipLevel.Friend)
+                                        civInfo.getEraObject().friendBonus[otherCiv.cityStateType.name]
+                                    else
+                                        civInfo.getEraObject().allyBonus[otherCiv.cityStateType.name]
+
+                if (relevantbonuses != null) {
+                    for (bonus in relevantbonuses) {
+                        if (bonus.getPlaceholderText() == "Provides [] Happiness") {
+                            if (statMap.containsKey("City-States"))
+                                statMap["City-States"] = statMap["City-States"]!! + bonus.getPlaceholderParameters()[0].toFloat()
+                            else
+                                statMap["City-States"] = bonus.getPlaceholderParameters()[0].toFloat()
+                        }
+                    }
+                } else {
+                    // Deprecated, assume Civ V values for compatibility
+                    if (otherCiv.cityStateType == CityStateType.Mercantile) {
+                        val happinessBonus = if(civInfo.getEraNumber() in 0..1) 2f else 3f
+                        if (statMap.containsKey("City-States"))
+                            statMap["City-States"] = statMap["City-States"]!! + happinessBonus
+                        else
+                            statMap["City-States"] = happinessBonus
+                    }
+                }
             }
         }